video: Add proper handling for ARGB8888
On some framebuffers (i.e MediaTek) the framebuffer comes with
the display layer (ovl0) set as fully transparent by default.
As such, writing the data with the fallback that was used, the
display would simply stay black. With this, we are setting the
pixel data to be fully opaque, which fixes the issue.

Signed-off-by: Igor Belwon <igor.belwon@mentallysanemainliners.org>